+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `Vector3Extensions.BestFitLine(this IList<Vector3> points)`
 - `Polygon.FromAlignedBoundingBox2d(IEnumerable<Vector3> points, Vector3 axis, double minSideSize = 0.1)`
 - `Transform.RotateAboutPoint` and `Transform.RotatedAboutPoint` convenience methods.
+- `Solid.ToCSG()` extension method is now an instance method on `Solid`.
 
 ### Changed